feat: add specific permissions feature to the user
