+++ src/controllers/userProfileController.js
@@ -63,7 +63,7 @@ const userProfileController = function (UserProfile) {
 
     UserProfile.find(
       {},
-      '_id firstName lastName role weeklyComittedHours email isActive reactivationDate createdDate endDate',
+      '_id firstName lastName role weeklyComittedHours email permissions isActive reactivationDate createdDate endDate',
     )
       .sort({
         lastName: 1,
@@ -169,6 +169,7 @@ const userProfileController = function (UserProfile) {
     up.collaborationPreference = req.body.collaborationPreference || '';
     up.timeZone = req.body.timeZone || 'America/Los_Angeles';
     up.location = req.body.location;
+    up.permissions = req.body.permissions;
 
     up.save()
       .then(() => {
@@ -275,9 +276,13 @@ const userProfileController = function (UserProfile) {
         record.weeklySummaryNotReq = req.body.weeklySummaryNotReq ? req.body.weeklySummaryNotReq : record.weeklySummaryNotReq;
         record.categoryTangibleHrs = req.body.categoryTangibleHrs ? req.body.categoryTangibleHrs : record.categoryTangibleHrs;
         record.totalTangibleHrs = req.body.totalTangibleHrs;
-        record.timeEntryEditHistory = req.body.timeEntryEditHistory;        
+        record.timeEntryEditHistory = req.body.timeEntryEditHistory;
         record.hoursByCategory = req.body.hoursByCategory;
         record.createdDate = moment(req.body.createdDate).toDate();
+
+        if (hasPermission(req.body.requestor.role, 'putUserProfilePermissions')) record.permissions = req.body.permissions;
+
+
         if (yearMonthDayDateValidator(req.body.endDate)) {
           record.endDate = moment(req.body.endDate).toDate();
           userData.endDate = record.endDate.toISOString();
