+++ src/models/role.js
@@ -6,5 +6,7 @@ const { Schema } = mongoose;
 const Role = new Schema({
   roleName: { type: String, required: true, unique: true },
   permissions: [String],
+  permissionsBackEnd: [String],
 });
+
 module.exports = mongoose.model('role', Role, 'role');
