+++ src/startup/db.js
@@ -2,6 +2,7 @@
 const mongoose = require('mongoose');
 const logger = require('./logger');
 const userProfile = require('../models/userProfile');
+const initialPermissions = require('../utilities/createInitialPermissions');
 
 mongoose.Promise = Promise;
 
@@ -14,6 +15,7 @@ const afterConnect = async () => {
       },
     );
 
+    await initialPermissions();
     if (!user) {
       userProfile.create({
         firstName: 'TimeArchiveAccount',
