+++ src/utilities/createInitialPermissions.js
@@ -1,0 +1,321 @@
+const Role = require('../models/role');
+
+const permissionsRoles = [
+  {
+    roleName: 'Administrator',
+    permissions: [
+      'seeWeeklySummaryReports',
+      'seeUserManagement',
+      'seeBadgeManagement',
+      'seePopupManagement',
+      'seeProjectManagement',
+      'seeTeamsManagement',
+      'deleteOwnBadge',
+      'modifyOwnBadgeAmount',
+      'assignBadgeOthers',
+      'editTimelogInfo',
+      'addTimeEntryOthers',
+      'deleteTimeEntryOthers',
+      'toggleTangibleTime',
+      'changeIntangibleTimeEntryDate',
+      'editTimeEntry',
+      'deleteTimeEntry',
+      'deleteWbs',
+      'addTask',
+      'deleteTask',
+      'editTask',
+      'addWbs',
+      'addProject',
+      'deleteProject',
+      'editProject',
+      'findUserInProject',
+      'assignUserInProject',
+      'unassignUserInProject',
+      'adminLinks',
+      'editUserProfile',
+      'assignTeamToUser',
+      'seeUserProfileInProjects',
+      'createTeam',
+      'editDeleteTeam',
+      'handleBlueSquare',
+      'resetPasswordOthers',
+      'dataIsTangibleTimelog',
+      'toggleSubmitForm',
+      'seePermissionsManagement',
+    ],
+    permissionsBackEnd: [
+      'seeBadges',
+      'assignBadges',
+      'createBadges',
+      'deleteBadges',
+      'updateBadges',
+      'createPopup',
+      'updatePopup',
+      'deleteProject',
+      'postProject',
+      'putProject',
+      'assignProjectToUsers',
+      'importTask',
+      'postTask',
+      'updateNum',
+      'updateTask',
+      'swapTask',
+      'deleteTeam',
+      'putTeam',
+      'assignTeamToUsers',
+      'editTimeEntry',
+      'deleteTimeEntry',
+      'updatePassword',
+      'getUserProfiles',
+      'getProjectMembers',
+      'postUserProfile',
+      'putUserProfile',
+      'putUserProfileImportantInfo',
+      'infringmentAuthorizer',
+      'deleteUserProfile',
+      'postWbs',
+      'deleteWbs',
+      'getAllInvInProjectWBS',
+      'postInvInProjectWBS',
+      'getAllInvInProject',
+      'postInvInProject',
+      'transferInvById',
+      'delInvById',
+      'unWasteInvById',
+      'getInvIdInfo',
+      'putInvById',
+      'getInvTypeById',
+      'putInvType',
+      'getAllInvType',
+      'postInvType',
+      'getWeeklySummaries',
+      'getTimeZoneAPIKey',
+      'checkLeadTeamOfXplus',
+    ],
+  },
+  {
+    roleName: 'Volunteer',
+    permissions: ['V'],
+    permissionsBackEnd: ['getReporteesLimitRoles'],
+  },
+  {
+    roleName: 'Core Team',
+    permissions: ['seeWeeklySummaryReports'],
+    permissionsBackEnd: [
+      'getUserProfiles',
+      'getProjectMembers',
+      'getAllInvInProjectWBS',
+      'postInvInProjectWBS',
+      'getAllInvInProject',
+      'postInvInProject',
+      'transferInvById',
+      'delInvById',
+      'unWasteInvById',
+      'getInvIdInfo',
+      'putInvById',
+      'getInvTypeById',
+      'putInvType',
+      'getAllInvType',
+      'postInvType',
+      'getWeeklySummaries',
+      'getTimeZoneAPIKey',
+      'checkLeadTeamOfXplus',
+    ],
+  },
+  {
+    roleName: 'Manager',
+    permissions: [
+      'seeWeeklySummaryReports',
+      'assignOnlyBlueSquares',
+      'editTask',
+    ],
+    permissionsBackEnd: [
+      'getUserProfiles',
+      'getProjectMembers',
+      'putUserProfile',
+      'infringmentAuthorizer',
+      'getReporteesLimitRoles',
+      'getAllInvInProjectWBS',
+      'postInvInProjectWBS',
+      'getAllInvInProject',
+      'postInvInProject',
+      'transferInvById',
+      'delInvById',
+      'unWasteInvById',
+      'getInvIdInfo',
+      'putInvById',
+      'getInvTypeById',
+      'putInvType',
+      'getAllInvType',
+      'postInvType',
+      'getWeeklySummaries',
+      'getTimeZoneAPIKey',
+      'checkLeadTeamOfXplus',
+    ],
+  },
+  {
+    roleName: 'Mentor',
+    permissions: [
+      'seeWeeklySummaryReports',
+      'assignOnlyBlueSquares',
+      'editTask',
+    ],
+    permissionsBackEnd: [
+      'getUserProfiles',
+      'getProjectMembers',
+      'putUserProfile',
+      'infringmentAuthorizer',
+      'getReporteesLimitRoles',
+      'getAllInvInProjectWBS',
+      'postInvInProjectWBS',
+      'getAllInvInProject',
+      'postInvInProject',
+      'transferInvById',
+      'delInvById',
+      'unWasteInvById',
+      'getInvIdInfo',
+      'putInvById',
+      'getInvTypeById',
+      'putInvType',
+      'getAllInvType',
+      'postInvType',
+      'getWeeklySummaries',
+      'getTimeZoneAPIKey',
+      'checkLeadTeamOfXplus',
+    ],
+  },
+  {
+    roleName: 'Owner',
+    permissions: [
+      'seeWeeklySummaryReports',
+      'seeUserManagement',
+      'seeBadgeManagement',
+      'seePopupManagement',
+      'seeProjectManagement',
+      'seeTeamsManagement',
+      'deleteOwnBadge',
+      'modifyOwnBadgeAmount',
+      'assignBadgeOthers',
+      'editTimelogInfo',
+      'addTimeEntryOthers',
+      'deleteTimeEntryOthers',
+      'toggleTangibleTime',
+      'changeIntangibleTimeEntryDate',
+      'editTimeEntry',
+      'deleteTimeEntry',
+      'deleteWbs',
+      'addTask',
+      'deleteTask',
+      'editTask',
+      'addWbs',
+      'addProject',
+      'deleteProject',
+      'editProject',
+      'findUserInProject',
+      'assignUserInProject',
+      'unassignUserInProject',
+      'adminLinks',
+      'editUserProfile',
+      'assignTeamToUser',
+      'createTeam',
+      'editDeleteTeam',
+      'seeUserProfileInProjects',
+      'handleBlueSquare',
+      'resetPasswordOthers',
+      'dataIsTangibleTimelog',
+      'addDeleteEditOwners',
+      'toggleSubmitForm',
+      'seePermissionsManagement',
+    ],
+    permissionsBackEnd: [
+      'seeBadges',
+      'assignBadges',
+      'createBadges',
+      'deleteBadges',
+      'updateBadges',
+      'createPopup',
+      'updatePopup',
+      'deleteProject',
+      'postProject',
+      'putProject',
+      'assignProjectToUsers',
+      'importTask',
+      'postTask',
+      'updateNum',
+      'updateTask',
+      'swapTask',
+      'deleteTeam',
+      'putTeam',
+      'assignTeamToUsers',
+      'editTimeEntry',
+      'deleteTimeEntry',
+      'updatePassword',
+      'getUserProfiles',
+      'getProjectMembers',
+      'postUserProfile',
+      'putUserProfile',
+      'putUserProfileImportantInfo',
+      'deleteUserProfile',
+      'infringmentAuthorizer',
+      'postWbs',
+      'deleteWbs',
+      'getAllInvInProjectWBS',
+      'postInvInProjectWBS',
+      'getAllInvInProject',
+      'postInvInProject',
+      'transferInvById',
+      'delInvById',
+      'unWasteInvById',
+      'getInvIdInfo',
+      'putInvById',
+      'getInvTypeById',
+      'putInvType',
+      'getAllInvType',
+      'postInvType',
+      'getWeeklySummaries',
+      'getTimeZoneAPIKey',
+      'checkLeadTeamOfXplus',
+    ],
+  },
+];
+const createInitialPermissionsFront = async () => {
+  permissionsRoles.forEach(
+    async ({ roleName, permissions, permissionsBackEnd }) => {
+      const allRoles = await Role.find();
+
+      // Add a new permission if the role Owner has been changed in the  permissionsRoles Array
+      if (roleName === 'Administrator') {
+        const ownerRoleDataBase = allRoles.find(
+          role => role.roleName === 'Administrator',
+        );
+        const { _id: roleId } = ownerRoleDataBase;
+        const permissionsBackOwnerDataBase = ownerRoleDataBase.permissionsBackEnd;
+        const permissionsFrontOwnerDataBase = ownerRoleDataBase.permissions;
+        const IsAllBackPermissionsOwnerUpdated = permissionsBackOwnerDataBase.every(perm => permissions.includes(perm));
+        const IsAllFrontPermissionsOwnerUpdated = permissionsFrontOwnerDataBase.every(perm => permissionsBackEnd.includes(perm));
+
+
+        if (!IsAllFrontPermissionsOwnerUpdated) {
+          await Role.findById(roleId, (_, record) => {
+            record.permissions = permissions;
+            record.save();
+          });
+        }
+        if (!IsAllBackPermissionsOwnerUpdated) {
+          await Role.findById(roleId, (_, record) => {
+            record.permissionsBackEnd = permissionsBackEnd;
+            record.save();
+          });
+        }
+      }
+      if (!allRoles.some(role => role.roleName === roleName)) {
+        const role = new Role();
+        role.roleName = roleName;
+        role.permissions = permissions;
+        role.permissionsBackEnd = permissionsBackEnd;
+        role.save();
+      }
+    },
+  );
+};
+module.exports = createInitialPermissionsFront;
